+++ src/components/main-screen/offers-sorting-form/offers-sorting-form.tsx
@@ -1,45 +1,12 @@
 import classNames from 'classnames';
 import { SortOptions } from '../../../const';
-import { useAppDispatch, useAppSelector } from '../../../hooks/redux';
-import { useEffect, useRef, useState } from 'react';
-import { setSortOptionAction } from '../../../store/action';
+import { useRef } from 'react';
+import { useOffersSortingForm } from './use-offers-sorting-form';
+
 
 function OffersSortingForm(): JSX.Element {
   const sortOptionsRef = useRef<HTMLFormElement>(null);
-  const activeSortOption = useAppSelector((state) => state.activeSortOption);
-  const [isOptionsOpened, setOptionsOpened] = useState(false);
-  const handleSortOptionsToggleClick = () => setOptionsOpened(!isOptionsOpened);
-
-  const dispatch = useAppDispatch();
-  const handleOptionClick = (sortOption: SortOptions) => {
-    dispatch(setSortOptionAction(sortOption));
-    setOptionsOpened(false);
-  };
-
-
-  const handleKeyDown = (evt: KeyboardEvent) => {
-    if (evt.key === 'Escape') {
-      setOptionsOpened(false);
-    }
-  };
-
-  const handleOutsideClick = (evt: MouseEvent) => {
-    if (sortOptionsRef.current?.contains(evt.target as Node) === false) {
-      setOptionsOpened(false);
-    }
-  };
-
-  useEffect(() => {
-    if (isOptionsOpened) {
-      document.addEventListener('keydown', handleKeyDown);
-      document.addEventListener('click', handleOutsideClick);
-
-      return () => {
-        document.removeEventListener('keydown', handleKeyDown);
-        document.removeEventListener('click', handleOutsideClick);
-      };
-    }
-  }, [isOptionsOpened]);
+  const {activeSortOption, isOptionsOpened, handleOptionClick, handleSortOptionsToggleClick} = useOffersSortingForm(sortOptionsRef);
 
   return (
     <form className="places__sorting" action="#" method="get" ref={sortOptionsRef}>
